+++ R/plot.R
@@ -1,7 +1,7 @@
 #' Plot Regression Coefficients
 #'
 #' The `plot()` function extracts `x$table_body` and passes the it to
-#' `GGally::ggcoef_plot()` along with a formatting options.
+#' `ggstats::ggcoef_plot()` along with a formatting options.
 #'
 #' \lifecycle{experimental}
 #' @param x 'tbl_regression' or 'tbl_uvregression' object
@@ -9,12 +9,12 @@
 #' for categorical variables. Default is `TRUE`
 #' @param remove_reference_rows logical indicating whether to remove reference rows
 #' for categorical variables. Default is `FALSE`.
-#' @param ... arguments passed to `GGally::ggcoef_plot(...)`
+#' @param ... arguments passed to `ggstats::ggcoef_plot(...)`
 #'
 #' @return a ggplot
 #' @name plot
 #'
-#' @examplesIf identical(Sys.getenv("IN_PKGDOWN"), "true") && broom.helpers::.assert_package("GGally", pkg_search = "gtsummary", boolean = TRUE)
+#' @examplesIf identical(Sys.getenv("IN_PKGDOWN"), "true") && broom.helpers::.assert_package("ggstats", pkg_search = "gtsummary", boolean = TRUE)
 #' glm(response ~ marker + grade, trial, family = binomial) %>%
 #'   tbl_regression(
 #'     add_estimate_to_reference_rows = TRUE,
@@ -29,7 +29,7 @@ plot.tbl_regression <- function(x,
                                 remove_header_rows = TRUE,
                                 remove_reference_rows = FALSE, ...) {
   check_dots_empty(error = function(e) inform(c(e$message, e$body)))
-  assert_package("GGally", fn = "plot.tbl_regression()")
+  assert_package("ggstats", fn = "plot.tbl_regression()")
 
   df_coefs <- x$table_body
   if (isTRUE(remove_header_rows)) {
@@ -40,7 +40,7 @@ plot.tbl_regression <- function(x,
   }
 
   df_coefs %>%
-    GGally::ggcoef_plot(exponentiate = x$inputs$exponentiate, ...)
+    ggstats::ggcoef_plot(exponentiate = x$inputs$exponentiate, ...)
 }
 
 #' @rdname plot